Introduced workaround for files needed by PSRAD.
* conrad/__init__.py:

Do not import psrad on top-level. Instead use lazy imports when it is run.
This is necessary, because PSRAD needs a lot of files to be in the current
working directory. The current workaround creates symlinks at runtime.

* conrad/conrad.py: Added a method to create an overview plot.

* conrad/plots.py: Added a function to plot profiles against height.

* conrad/psrad/psrad.py: Reverted to original version.

* conrad/utils.py:

Added a helper class which creates symlinks to files needed by psrad.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10016 0c488439-3f28-0410-89a5-82d5fbb9eead