Added first example studies as sanity check.
* conrad/atmosphere.py (Atmosphere.set():

Added a method to set the values for a variable in an Atmosphere. It is
possible to pass single values or arrays.

* scripts/isothermal_rce.py:

Test simulations with isothermal atmospheres covering a wide range of
temperatures.

* scripts/temperautre_shift_rce.py:

Simulations with shifted temperature profiles covering a wide range of
temperatures.

* scripts/isothermal_comparison.ipynb:

Visualisations for the isothermal and shifted temperature experiments.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10056 0c488439-3f28-0410-89a5-82d5fbb9eead