Added a convection scheme that is still under **massive** development.
* conrad/atmosphere.py (AtmosphereConvective):

Added new methods `find_first_unstable_layer` and
`convective_adjustment` to the Atmosphere base class.

Implemented an AtmosphereConvective that make use of the new methods
when implementing the required `adjust` method.

* conrad/surface.py (SurfaceCoupled):

Utility class that couples the surface temperature and pressure to the
lowest atmosphere level.

* conrad/utils.py (create_relative_humidity_profile):

Utility function to create a relative humidity profile with an
exponential decay with increasing heights.

* scripts/test_rce.py: Adjusted test script to current development.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10063 0c488439-3f28-0410-89a5-82d5fbb9eead