* atmosphere.py:
* (Atmosphere):

Implemented a new property `relative_humidity` which allows to return
and set the relative humidity conveniently. The method internally works
with pressure, temperature and VMR.

* (AtmosphereConvective):

The convection scheme does no longer start at the bottom but the first
convectively unstable layer found in the temperature profile.

* surface.py (Surface):

When initializing a Surface using an Atmosphere object, the bottom layer
is no longer copied completely. The pressure is extrapolated to prevent
a halflayer with zero thickness between the lowest atmosphere and the
surface.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10085 0c488439-3f28-0410-89a5-82d5fbb9eead