* utils.py:
* (append_timestep_netcdf):

Improved if-clause which determines if to append a variable. It now
simply checks for a `time` dimension and if the variable itself is not
a dimension.

* (calculate_halflevels):

New function to interpolate halflevel values of a given variable.

* constants.py: Added description for `phlev`.

* atmosphere.py (Atmosphere):

Added air pressure at halflevel to the list of available coordinates.

* (Atmosphere.append_description):

Check if description is actually present. `None` attribtues seem to
break xarray objects.

* radiation/__init__.py:

All fluxes are now stored on proper halflevel coordinates.

git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10122 0c488439-3f28-0410-89a5-82d5fbb9eead