Major API change to the Surface class!
* conrad/surface.py:

* (Surface.adjust):

The `adjust` method no longer uses heatingrates as input! Instead the
net solar fluxes and downward lonwave fluxes are passed. They are used
to calculate the resulting surface temperature.

* (SurfaceCoupled):

The surface model coupled to the atmosphere has been removed. It is no
longer needed as the new implemented adjustment method couples the
surface to the atmosphere implicitly.

* conrad/radiation/__init__.py:

All shortwave fluxes and heatingrates are divided by a factor of two.
This is an very straight-forward workaround to take a day and night time
fo the day into account.

* conrad/constants.py: Added the Stefan-Boltzmann constant.

* conrad/conrad.py (RCE.run): Adapted the call of `surface.adjust`.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10154 0c488439-3f28-0410-89a5-82d5fbb9eead