Major change of the Surface model signature!
* conrad/surface.py:

* Surface: The `Surface` class now inherits from `xarray.Datasets`.

This change allows to set properly describe all attributes. The surface
parameters are now also written to the output netCDF file.

* SurfaceNoHeatCapacity: Renamed from `SurfaceAdjustableTemperature`.

* SurfaceHeatCapacity: New model describing a sufrace with heat capacity.

This model simulates surfaces that adjust to incoming fluxes with a
delay.  This allows to keep the differences in radiative fluxes small.
Also a possible adjustment of the relative humidity in lower atmospheric
layers is more moderate.

* conrad/conrad.py:

* create_netcdf, append_to_netcdf: Include surface parameters.

* run: Adapted new signature of `Surface.adjust()`.

* conrad/constants.py: Inclued descriptions for surface parameters.

* conrad/utils.py (append_description):

Function to append variable descriptions to Datasets (e.g. Atmosphere).
This replaces the same-named methods in all model classes.

* conrad/atmosphere.py, conrad/radiation/__init__.py:

Use `append_description()`.`


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10179 0c488439-3f28-0410-89a5-82d5fbb9eead