Added support for float and array-like lapse rate attributes.
* atmosphere.py (Atmosphere):

* (__init__):

The air temperature lapse rate can be passed as float or a array. The
different cases are recognized and `xr.Dataset` dimensions are set
accordingly.

* (Atmosphere.from_atm_fields_compat,
   Atmosphere.from_dict,
   Atmosphere.from_netcdf):

All classmethods are collecting additional keyword arguments and pass
them to the constructor. This allows, for example, to set the lapse rate
when creating atmosphere models from external data.

* constants.py: Added description for air temperature lapse rate.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10205 0c488439-3f28-0410-89a5-82d5fbb9eead