Added an optional diurnal cycle for the solar angle.
* conrad/radiation/__init__.py (Radiation.adjust_solar_angle):

All radiation models offer an optional diurnal cycle for the solar
angle. This behavior can be controlled through the `diurnal_cycle`
keyword and is turned off by default.

If turned on, the solar angle oscillates sinusodila between 90 deg and
the given zenith angle.

* conrad/conrad.py:

Include the newly implemented `Radiation.adjust_solar_angle()` to the
model control flow.

Improved inline comments.

* conrad/atmosphere.py (Atmosphere.apply_H20_limits):

Implement the functionality to set an upper limit for water vapor VMRs.
This has currently no effect, as the threshold is set to `1.0`.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10213 0c488439-3f28-0410-89a5-82d5fbb9eead