Added a method to calculate geopotential height.
* conrad/atmosphere.py (Atmosphere.calculate_height):

Method to calculate geopotential height for current atmospheric state.
The values are stored as 'z' in the atmosphere Dataset, they replace the
old fixed height coordinates.

The method shall be called after each adjustment of the atmosphere. This
is done for all classmethod as well as all adjust methods that are
currently available.

* conrad/constants.py: Updated description of (geopotential) height.

* conrad/surface.py: Cleaned comments.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10227 0c488439-3f28-0410-89a5-82d5fbb9eead