*atmosphere.py: AtmosphereConvective
Added the function readjust, to preform an adjustment to the convective adjustment for cases when it is not very energy conserving. This removes the wavey pattern for the moist adiabatic lapse rate. This readjustment can be applied iteratively until the convective adjustment is sufficiently close to being energy conserving.

Output convective_top also for the fixed surface temperature case.



git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10236 0c488439-3f28-0410-89a5-82d5fbb9eead