*atmosphere.py
Updated the convective adjustment, so that the we integrate over the lapse rate as calculated on the half pressure levels. This means that the surface temperature is no longer the same as the lowest atmospheric level, but is at a pressure height equal to the lowest half pressure level and has a temperature which follows the chosen lapse rate.

Added comments.



git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10308 0c488439-3f28-0410-89a5-82d5fbb9eead