Refactored the ozone_profile function.
* conrad/data/ozone_profile.nc:

Moved the data from the scripts directory into the package itself.
This clarifies that it is needed for internal functions.

* conrad/utils.py (ozone_profile):

Converted all positional arguments into keyword arguments. This allows
to call the function without arguments at all. Default behavior is to
simply return the ozone profile stored in the included data file.


git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10359 0c488439-3f28-0410-89a5-82d5fbb9eead