Simplify the handling of surface pressure.
* conrad.py (run):

Initialize the surface pressure to be at the lowest atmospheric
half-level pressure. This ensures handling consistent with PSrad.

Do not store the initial relative humidity profile. This is outdated
since the introduction of selfconatined humidty handlers.

* surface.py (Surface.__init__): Remove `pressure` keyword.

Do not allow the user to set the surface pressure directly. This is done
automatically to prevent inconsistencies.

git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10406 0c488439-3f28-0410-89a5-82d5fbb9eead