*atmosphere.py
calculate_height: Changed the calculation of z (height), so that the difference between the surface and the first atmospheric pressure level is more consistent with other quantities.

adjust method: Moved the calculation of lapse rate before the radiative heating rates are applied.

*lapserate.py

get: made this a function of the atmosphere.

Tried to implement a new method of calculating the moist adiabat, currently unsucessfully (surface issues).




git-svn-id: https://arts.mi.uni-hamburg.de/svn/internal/users/lkluft/conrad/trunk@10413 0c488439-3f28-0410-89a5-82d5fbb9eead