Include latent heat change in convection
Take into account the energy change due to water vapour changes
from the previous timestep to this timestep. This does not
consider the energy required to loft the water vapour from the
surface upwards into the atmosphere. It assumes a constant
value of latent heat (independent of temperature and pressure).