Throw error if no norm level for coupled upwelling
This should avoid problems if, eg. the user does not specify a
normalisation level and there is initially no convection and when
convection starts it does not reach the top of the troposphere.
Such a scenario may arise if the initial temperature profile is
unrealistic or an applied perturbation acts to cool the surface.