Move specification of diurnal cycle to core
The radiation submodel does not include a time dimension, and the
instance initialisation does not require the time to be specified,
so it does not make much sence to specify the diurnal cycle there.
This can lead to misunderstandings, for example if the radiation
is used independently of konrad for flux calculations.
Nevertheless the current solar angle is still stored inside the
radiation class and can be updated by passing a time to the
'adjust_solar_angle' function.